refactored to decentralize axios request
