+++ client/src/SharedComponents/reviewMeta.js
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+const currentMetaReview = atom({
+  key: 'currentMetaReview',
+  default: {},
+});
+
+export default currentMetaReview;
